scripts/legacy/requirements.txt
qdrant-client>=1.12.0
fastembed>=0.4.0
python-dotenv>=1.0.0
